requirements.txt
# 核心依赖（固定兼容版本）
Flask==2.2.5
Werkzeug==2.3.7
Flask-WTF==1.2.1
gunicorn==21.2.0
markupsafe==2.1.3

# 其他依赖
alibabacloud-dysmsapi20170525==4.1.2
aliyun-python-sdk-dysmsapi==2.1.1
blinker==1.8.2
flask-login==0.6.3
flask-sqlalchemy==3.0.3
email-validator==1.1.3
python-dotenv==0.19.0
requests==2.32.4
idna==3.10
wtforms==3.1.2
cryptography==44.0.3

# 数据库相关
SQLAlchemy==2.0.25